chore: rename OtherWalletPage to OtherWalletLoginPage
